python-client/wmill_pg/pyproject.toml
[tool.poetry]
name = "wmill-pg"
version = "1.5.0"
description = "An extension client for the wmill client library focused on pg"
license = "Apache-2.0"
homepage = "https://windmill.dev"
documentation = "https://docs.windmill.dev"

authors = ["Ruben Fiszel <ruben@rubenfiszel.com>"]

readme = "README.md"
packages = [
    {include = "wmill_pg"},
]
include = ["wmill_pg/py.typed"]

[tool.poetry.dependencies]
python = "^3.7"
wmill = "^1.5.0"
psycopg2-binary = "*"

[build-system]
requires = ["poetry>=1.0.2", "poetry-dynamic-versioning"]
build-backend = "poetry.masonry.api"

[tool.poetry-dynamic-versioning]
enable = true
vcs = "git"
style = "semver"
format-jinja = "{% if distance == 0 %}{{ base }}{% else %}{{ base }}+{{ distance }}.{{ commit }}{% endif %}"


[tool.black]
line-length = 120
target_version = ['py36', 'py37', 'py38']
exclude = '''
(
  /(
    | \.git
    | \.venv
    | \.mypy_cache
  )/
)
'''

[tool.isort]
line_length = 120
profile = "black"
